requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.33
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Werkzeug==3.0.1
zipp==3.8.0
